Refactoring + display entire level with scaling

+++ src/engine/graphics/MarioImage.java
@@ -16,8 +16,8 @@ public class MarioImage extends MarioGraphics {
     public void render(Graphics og, int x, int y) {
         if (!visible) return;
 
-        int xPixel = (int) x - originX;
-        int yPixel = (int) y - originY;
+        int xPixel = x - originX;
+        int yPixel = y - originY;
         Image image = this.sheet[index % sheet.length][index / sheet.length];
 
         og.drawImage(image, xPixel + (flipX ? width : 0), yPixel + (flipY ? height : 0), flipX ? -width : width, flipY ? -height : height, null);
